Fix ABA test when remote is not available

+++ gemma-core/src/test/java/ubic/gemma/core/image/aba/AllenBrainAtlasServiceTest.java
@@ -23,13 +23,14 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.springframework.beans.factory.annotation.Autowired;
-import ubic.gemma.persistence.service.genome.gene.GeneService;
 import ubic.gemma.core.util.test.BaseSpringContextTest;
 import ubic.gemma.core.util.test.category.AllenBrainAtlasTest;
 import ubic.gemma.model.genome.Gene;
+import ubic.gemma.persistence.service.genome.gene.GeneService;
 import ubic.gemma.persistence.service.genome.taxon.TaxonService;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assume.assumeNotNull;
 import static ubic.gemma.core.util.test.Assumptions.assumeThatResourceIsAvailable;
 
@@ -66,7 +67,9 @@ public class AllenBrainAtlasServiceTest extends BaseSpringContextTest {
 
     @After
     public void Cleanup() {
-        this.geneService.remove( gene );
+        if ( gene != null ) {
+            this.geneService.remove( gene );
+        }
     }
 
     @Test
